now configue checks that Fortran77 compiler works
So, for the intricacies of autoconf, the removal of the macro `AC_PROG_CC'
from configure.ac had the side effect of making the macro `AC_PROG_F77'
emit shell code checking that the fortran compiler can produce working
executables.

This check was not done before (I have no idea why), and unfortunately it
breaks the recommended way of disabling the use of the fortran compiler
in the testsuite -- i.e. passing F77=NONE to the configure script.

On the other hand, these new checks done on the selected fortran compiler
are quite useful, and we want to keep them enabled.

So we simply disable the call to AC_PROG_F77 if the $F77 variable is set
to NONE.

M: m4/ratfor_prog_f77.m4:
  * in `RAT4_PROG_F77': do not look for and check the Fortran 77 compiler
    (i.e. do not call `AC_PROG_F77') if $F77 is set to "NONE".

M: NEWS:
  * tell that configure now check that the fortran compiler found (or
    specified by the user) works as expected (in particular, can create
    simple executables)

M: configure.ac:
  * moved the checks for ratfor after the inclusion of the file
    `maint/config.site'