do not automatically generate list of test scripts
OK, let's face it: the automatic generation of the list of tests was
stupid, complicated, and WRONG.  The list of test scripts should be
have been regenerated whenever a test script was added in the `tests/'
directory, but in truth THERE IS NO EASY NATURAL WAY FOR MAKE TO
AUTOMATICALLY DISCOVER THIS ADDITION.  And we didn't divise and code
an hack to accompish this.  So, when a test script was addded, the
list was not updated -- in fact, nothing was updated, and nothing
happened.  D'OH!

So let's revert to the old behaviour of explicitly writing the list of
tests, and checking it through a comparison with the list of `*.test'
files contentained in the `tests/' directory.  It is simpler, more
solid, and it *works* (and it can nicely reuse code from SteLib, too).

A: tests/CheckTestsList.am:
  * added from SteLib

M: tests/Makefile.am:
  * re-inserted hard-wired definition of $(ALL_TESTS)
  * do not include `TestsList.am' anymore
  * include `CheckTestsList.am'
  * removed references to `TestsList.am or `gen-tests-list.sh', and
    any rule linked to them

D: build-aux/gen-tests-list.sh:
D: bootstrap.sh:
  * removed: no more needed

M: HACKING:
  * adjusted accordingly: do not refer to bootstrap.sh anymore

M: .gitignore:
  * removed references to `tests/TestLisat.am'