new target maintainer-check (check list of tests moved there)
* Makefile.am (maintainer-check-list-of-tests): New target,
delegating to the equally-named target in `tests/Makefile.am'.
(maintainer-check): New target.  Depends from the just-described
new target `maintainer-check-list-of-tests'.
* tests/Makefile.am (check-local): It doesn't depend anymore on
`check-list-of-tests'. Thus removed, as it's not useful anymore.
(ALL_TESTS): Deleted, no more needed.  Use $(TESTS) directly.
(TESTS): Directly define to the list of test scripts.
* sanity-check/Makefile.am: Likewise.
* lib/CheckTestsList.am: Updated from SteLib.
(check-list-of-tests): Target deleted, renamed to ...
(maintainer-check-list-of-tests): ... this.
It now checks the contents of the `$(TESTS)' variable, not of the
`$(ALL_TESTS)' variable.
It consider also the test scripts in `$(builddir)', not only
those in `$(srcdir)' (this can account for generated test scripts).
It does not require anymore the definition of the $(DIFF_U)
variable.
Finally, the verbosity of the target has been decreased.