ChangeLog generation: normal target, not a dist-hook anymore
* lib/GitGenerateChangeLog.am: Updated from SteLib.
In this new version of the file, the code providing generation of
the ChangeLog is moved back to a "normal rule", rather than being
left in a dist-hook.  The good news is that this time, differently
from our previous attempt, we provide all the necessary bells and
whistles, so that the rule:
  - works even with non-GNU make;
  - is strict and picky if maintainer rules are enabled;
  - is lax enough otherwise (e.g. it does not require the source
    tree to have a full git repository), so that "make distcheck"
    continue to work.
  - is generally more configurable;
A more detailed list of changes follows.
(dist-hook-generate-changelog): Rule renamed as...
(git-changelog-gen): ... this new phony target, which generates the
ChangeLog in the current $(builddir) rather than in the $(distdir).
It also uses ...
(FORCE_CHANGELOG_GENERATION, changelog_start_date,
gitlog_to_changelog_extra_options): ... these user-overridable
variables (undefined in this file, might be set to desired defaults
in the including `Makefile.am' file).
($(top_srcdir)/ChangeLog): New target, depends on it.
(dist-hook): Don't depend on `dist-hook-generate-changelog' rule
anymore.
(maintainer-clean-local): Extend with ...
(git-changelog-gen-maintainerclean): ... this rule, removing the
generated ChangeLog file.
(mostlyclean-local): Extend with ...
(git-changelog-gen-mostlyclean): ... this rule, removing temporary
file(s) used in the generation of the ChangeLog file.
(#heading-comments): Updated accordingly to the changed semantics.
* Makefile.am: Some comments properly fixed.
(EXTRA_DIST): Added `$(top_srcdir)/ChangeLog'.
* .gitignore: Added `ChangeLog' again.