Testsuite: work around a bug in FreeBSD make w.r.t. parallel tests
* tests/graybox/Makefile.am: Work around a bug in FreeBSD make
w.r.t. parallel tests: failures were not properly reported in
exit status of make when it ran more tests in parallel.
(check-workaround-freebsd-make-bug): New target.
(check-local): Depend on it.
* tests/blackbox: Submodule updated from commit `95b259b564fea...'
to commit `039afe76280fa...'.  This new version fixes a bug in the
subpackage's version of SteLib's file `MostlyCleanConfigureTmp.am',
and add the same workaround for the bug of FreeBSD make described
above.