Keep only text of GPLv3, not also of LGPLv3
We kept also a copy of the LGPLv3, since some files copied verbatim
from SteLib are licensed under it.  However, this seems excessive,
since those files contains a proper copyright notice (and a link
from where the LGPL text can be easily obtained), while the bulk of
the Ratfor Testsuite is covered by the GPL (v3 o later).  So from
now on we'll just keep a copy of the GPLv3.

* GPLv3: File deleted, renamed to ...
* COPYING: .. this.
* LGPLv3: File deleted.
* Makefile.am (EXTRA_DIST): Removed GPLv3 and LGPLv3.