Bugfix in Makefile.am: define $(TEST_SUBDIRS)
* Makefile.am (TEST_SUBDIRS): New variable, list of directories
where to recurse into for testing and related maintainer-checks.
This variables was already used as the list of a `for' loop in
the rules of `maintainer-check-list-of-tests' target; in fact,
according to the logs, the $(TEST_SUBDIRS) variable *should* have
been introduced in commit `f9a3f21261b9ca1cf0a6317df94e67a3...'.
However, probably for an oversight, it wasn't truly set in that
commit, this leaving the "for loop's list" empty.  With most
shells (e.g. bash and zsh), this didn't caused a syntax error,
but (worse) it made the shell fail to execute the loop's body
-- so that no check was really run, and there was no way to easily
notice that.  Fortunately, Heirloom Sh and/or Solaris Sh choke on
an a `for' loop with an empty list, and this let us uncover and
fix the bug.