Improve and extend tests on ratfor comments.
* tests/comments-keep_f77.test: Improved and extended.
* tests/comments-keep.test: File deleted, renamed to ...
* tests/comments-keep-at-stmt-begin.test: ... this, and extended
to check that ratfor comments are kept if ratfor is called with
the `-C' switch, and they are placed at the beginning of a
statement.
* tests/comments-basic.test: File deleted, split into to ...
* tests/comments-not-in-strings.test: ... this, checking that
ratfor comments are kept when placed inside a string ...
* tests/comments-discard-at-stmt-begin.test: ... and this,
checking that comments are by default discarded if placed at
the beginning of a statement ...
* tests/comments-discard-in-stmt-middle.test: ... and this,
checking that comments are by default discarded if placed in
the middle of a statement ...
* tests/comments-discard-after-numerical-label.test: ... and
this, checking that comments are by default discarded if
placed after a numberical label.
* tests/comments-at-stmt-begin.data: New data file, used by
test scripts `comments-discard-at-stmt-begin.test' and
`comments-keep-at-stmt-begin.test'.
* tests/Makefile (TESTS): Updated accordingly.
(comments-discard-at-stmt-begin.log,
comments-keep-at-stmt-begin.log): Depend on file
`comments-at-stmt-begin.data'.
(EXTRA_DIST): Added `comments-at-stmt-begin.data'.