Dumbing down unit tests: don't use libcheck
While experimenting and writing the first unit tests, it turned out
that the use of the `check' library is truly an overkill, and adds
layers of useless complexity.  So we are removing support for it,
and planning to write unit tests by incrementally writing simple
custom macros/functions when the need arise.

* configure.ac (AX_CHECK_LIBCHECK): Call to ths macro removed.
* m4/ax_check_libcheck.m4: File removed, no more needed.