Refactor and fix code for macro definition/expansion
* src/define.c (macro_definition_lookup): New saner interface:
instead of returning a boolean indicating if the given macro is
defined, and if so copying its value in a user-provided buffer,
simply return a pointer to the string already saved in the hash
containing the list of defined macros (return NULL if the macro
is undefined).  This also avoid a nasty use of the infamous
`strcpy' function.
* src/define.c (macro_definition_lookup): Prototype modified
accordingly.
* src/tokenizer.c (get_expanded_token): Call to that function
modified accordingly.  Nname and type ofa local variable changed:
char defn[] => char *macro_body.