Check ratfor keywords in statement middle (with f77)
* tests/if-in-statement-middle_f77.test: New test.  Checks that
the ratfor `if' keyword is not not recognized in the middle of
a statement.
* tests/else-in-statement-middle_f77.test: New test.  Checks
that the ratfor `else' keyword is not recognized in the middle
of a statement.
* tests/while-in-statement-middle_f77.test: New test.  Checks
that the ratfor `while' keyword is not recognized in the middle
of a statement.
* tests/for-in-statement-middle_f77.test: New test.  Checks that
the ratfor `for' keyword is not recognized in the middle of a
statement.
* tests/repeat-in-statement-middle_f77.test: New test.  Checks
that the ratfor `repeat' keyword is not recognized in the middle
of a statement.
* tests/until-in-statement-middle_f77.test: New test.  Checks
that the ratfor `until' keyword is not recognized in the middle
of a statement.
* tests/do-in-statement-middle_f77.test: New test.  Checks that
the ratfor `do' keyword is not recognized in the middle of a
statement.
* tests/case-in-statement-middle_f77.test: New test.  Checks that
the ratfor `case' keyword is not recognized in the middle of a
statement.
* tests/default-in-statement-middle_f77.test: New test.  Checks
that the ratfor `default' keyword is not not recognized in the
middle of a statement.
* tests/switch-in-statement-middle_f77.test: New test.  Checks
that the ratfor `switch' keyword is not recognized in the middle
of a statement.
* tests/keyword-in-statement-middle_f77.sh: Helper script used
by the previously listed new tests.
* tests/Makefile.am (TESTS): Updated accordingly.
(EXTRA_DIST): Added `keyword-in-statement-middle.sh'.
(repeat-in-statement-middle_f77.log): Depends from file
`keyword-in-statement-middle.sh'.
(until-in-statement-middle_f77.log): Likewise.
(switch-in-statement-middle_f77.log): Likewise.
(case-in-statement-middle_f77.log): Likewise.
(default-in-statement-middle_f77.log): Likewise.
(else-in-statement-middle_f77.log): Likewise.
(if-in-statement-middle_f77.log): Likewise.
(while-in-statement-middle_f77.log): Likewise.
(for-in-statement-middle_f77.log): Likewise.
(do-in-statement-middle_f77.log): Likewise.