Keywords `break' and `next' in stmt middle: check that are not active
* tests/break-in-statement-middle.test: New test.  Check (with
grep) that the ratfor `break' keyword is not not recognized in
the middle of a statement.
* tests/break-in-statement-middle_f77.test: New test.  Check (with
f77) that the ratfor `break' keyword is not not recognized in the
middle of a statement.
* tests/next-in-statement-middle.test: New test.  Check (with
grep) that the ratfor `break' keyword is not not recognized in
the middle of a statement.
* tests/next-in-statement-middle_f77.test: New test. Check (with
f77) that the ratfor `break' keyword is not not recognized in the
middle of a statement.
* tests/Makefile.am (TESTS): Updated accordingly.
(next-in-statement-middle.log): Depends from file
`keyword-in-statement-middle.sh'.
(break-in-statement-middle.log): Likewise.
(next-in-statement-middle_f77.log): Depends from file
`keyword-in-statement-middle_f77.sh'.
(break-in-statement-middle_f77.log): Likewise.