Keywords `break' and `next' after label numbers: check that they work
* tests/break-after-numeric-label.test: New test.  Check (with
grep) that the ratfor `break' keyword is not not recognized in
the middle of a statement.
* tests/break-after-numeric-label_f77.test: New test.  Check (with
f77) that the ratfor `break' keyword is not not recognized in the
middle of a statement.
* tests/next-after-numeric-label.test: New test.  Check (with
grep) that the ratfor `break' keyword is not not recognized in
the middle of a statement.
* tests/next-after-numeric-label_f77.test: New test. Check (with
f77) that the ratfor `break' keyword is not not recognized in the
middle of a statement.
* tests/control-structure-after-numeric-label_f77.sh: Fix a bug
preventing the test on `break' from working.
* tests/Makefile.am (TESTS): Updated accordingly.
(next-after-numeric-label.log): Depends from file
`keyword-after-numeric-label.sh'.
(break-after-numeric-label.log): Likewise.
(next-after-numeric-label_f77.log): Depends from file
`keyword-after-numeric-label_f77.sh'.
(break-after-numeric-label_f77.log): Likewise.