Update testlib.sh from SteLib (fix bug when F77=fc)
* lib/testslib.sh: Updated from SteLib.  This new version ensure
that the `run_command' subroutine does not execute shell builtins
but only external commands, even under the most "restive" Korn
shells (like pdksh).  This fixes a bug in the test scripts, which
occurred a fortran compiler named `fc' was used, since it would
collide with the `fc' builtin of those korn shells.