New maintcheck: fortran compiler requested correctly (if and only if needed)
* am/MaintainerChecks.am (sc--requesting-of-fortran-compiler): New
maintainer check, verifying that tests expecting to use the fortran
compiler correctly check whether it is available, and that tests
not expecting to use the fortran compiler do not check wheter it is
available.
* build-aux/requesting-of-fortran-compiler.pl: New script,
implements the `sc--requesting-of-fortran-compiler' target.
* tests/control-structure-with-empty-body_f77.sh: Bugfix: added
call to `rat4t_require_fortran_compiler', as suggested by new
maintainer checks.