New graybox test (ratfor doesn't hang on comments)
* tests/graybox/comments-no-hang.test: New test, checking that
C Ratfor doesn't hang if the last line of input is a comment
(this already happened in practice, so this is in fact a sort
of regression test).  This test script is also in the blackbox
tests offered by the Ratfor Testsuite package, but it's going
to be removed from there, as it is way too much specific to
C Ratfor.
* tests/graybox/Makefile.am (TESTS): Extended accordingly.