Tokens after "until" don't trigger errors anymore
* src/parser.c (parse): Do not trigger a syntax error if tokens
are found after an "until" statement; instead, consider "until"
statements as truly "auto-terminating".
Other very minor cosmetic changes & fixes.
* src/graybox/tokens-after-until.test: Test this new behaviour.
* src/graybox/tokens-after-until_f77.test: Likewise.
* src/graybox/Makefile.am (TESTS): Updated accordingly.
* xfail-tests: Extended accordingly.
(until-with-extra-traling-tokens-fail.test): New xfailing test,
it has become obsoleted.
(comments-no-extra-white-spaces.test): The changes in parser have
revealed a limitation in the code generation subroutines of C
Ratfor, which causes this test to fail.  Once we'll have fixed
those limitations, this test should pass again, but for the moment
it's simpler to mark it as xfailing.