Invert semantic of a test on repeat/until
In the previous versions of The Ratfor Testsuite, we checked
that the tested ratfor preprocessor issued an error when finding
a "non-statement-terminating" token placed right after the
"until(COND)" part in a "repat/until" statement.

For example, we expected that an input like:
  repeat { stmt1 } until(cond) stmt2
caused ratfor to issue an error message like:
  "extra tokens after until."

However, during the developement of C Ratfor, it turned out that
the most proper implementation of the "repeat/until" statement
_required_ any "until(COND)" to be a "statement terminator" by
itself.

Given that, the most natural thing to do when encountering an
input like:
  repeat { stmt1 } until(cond) stmt2
is to expect it to be (loosely) equivalent to:
  repeat { stmt1 } until(cond)
  stmt2
This is also the approach taken by one of the (probably) most
long-lived, highest-quality implementation of ratfor, `rf' by
M.J Harrision, written in ratfor itself (and obviously able to
auto-bootstrap).

This new expected behaviour requires that we to _reverse_ the
logic of the test(s) checking how ratfor behaves when encountering
non-statement-terminating tokens after an "until(COND)".

* tests/until-with-extra-traling-tokens-fail.test: Test deleted.
* tests/until-without-statement-terminator.test: New test.
* tests/until-without-statement-terminator_f77.test: Likewise.
* tests/until-nested-2.test: Fixed.
* tests/Makefile.am (handwritten_tests): Updated accordingly.