Fix NASTY #-related bug in testscripts generation.
There's simply no way to save the `#' character in a Makefile
variable portably.  Our attempt to do so was botched and broken,
and didn't work with at least Solaris make (but it didn't work
in a subtle, non-evident way, thus going undetected for quite a
long time).  Fortunately, we have a workaround.

* tests/Makefile.am (sharp): Variable removed, no more needed (and
its definition was unportable anyway).
(autogen_sh__pre): Use a `tr'-based hack to put a literl `#'
character in output files.
(autogen_sh__post): Do not append a vim modeline to generated
files anymore.
* NEWS: Updated.